Fix review findings: bounds checking, type safety, and overflow issues (#77)
* Initial plan

* Address PR review comments

- Fix ImageRGB.getSubimage() bounds checking to prevent out-of-bounds access
- Add validation to ImageRGB.alphaBlendOver() for matching dimensions
- Fix SixelDecoder.resizeImage() to handle growing beyond current size
- Simplify byte-to-int conversion in ECMA48.java using standard bitwise AND
- Fix null safety checks in Cell.compareCellImages()
- Remove useless null check in Cell.getImage()
- Fix typo in MathUtils javadoc - "mathemathical" → "mathematical"
- Add missing javadoc for parameters V, d, e in MathUtils methods
- Fix redundant condition check in ECMA48Terminal
- Add equals() method to ColorIdx class (overrides hashCode)
- Fix incomparable type comparisons in TImageWindow (TKeypressEvent vs TKeypress)
- Fix incomparable type comparison in TImage (TCommandEvent vs TCommand)
- Fix potential integer overflow in HQSixelEncoder color calculations


* Apply suggestions from code review

* Apply review feedback: allow larger subimages and simplify

- Remove equals() method from ColorIdx (hashCode-only is acceptable here)
- Change ImageRGB.getSubimage() to allow requesting subimages larger than source
- Add conditional check to only copy pixels that exist within source bounds
- Simplify SixelDecoder.resizeImage() to just call getSubimage directly


* Update code/src/main/java/casciian/bits/Cell.java


---------

+++ code/src/main/java/casciian/backend/HQSixelEncoder.java
@@ -407,9 +407,9 @@ public class HQSixelEncoder implements SixelEncoder {
                     int red   = (rgb >>> 16) & 0xFF;
                     int green = (rgb >>>  8) & 0xFF;
                     int blue  =  rgb         & 0xFF;
-                    totalRed   += color.count * red;
-                    totalGreen += color.count * green;
-                    totalBlue  += color.count * blue;
+                    totalRed   += (long) color.count * red;
+                    totalGreen += (long) color.count * green;
+                    totalBlue  += (long) color.count * blue;
                     count += color.count;
                 }
                 if (count == 0) {
